- Added functions to create snake bodies, copy them, and add then to a linked list - Split Snake_initialization into c and h files
